+++ CHANGELOG.md
@@ -4,9 +4,14 @@ Here we will note all the changes made to the framework.
 
 ## Version 2.1.4
 
+🌟 Improvements:
+
+- Bump versions.
+
 🔧 Bugfix:
 
 - Fix an issue with the fixed size number generation.
+- Fix an issue with the extension ladder.
 
 ## Version 2.1.3